fix int to double

+++ app/src/main/java/com/phpsysinfo/xml/PSINetworkInterface.java
@@ -5,11 +5,11 @@ public class PSINetworkInterface {
 	public String name = "";
 	public double rxBytes = 0;
 	public double txBytes = 0;
-	public int err = 0;
-	public int drops = 0;
+	public double err = 0;
+	public double drops = 0;
 	public String info = "";
 
-	public PSINetworkInterface(String name, double rxBytes, int txBytes, int err, int drops, String info) {
+	public PSINetworkInterface(String name, double rxBytes, double txBytes, double err, double drops, String info) {
 		this.name = name;
 		this.rxBytes = rxBytes;
 		this.txBytes = txBytes;
@@ -36,13 +36,13 @@ public class PSINetworkInterface {
 	public void setTxBytes(double txBytes) {
 		this.txBytes = txBytes;
 	}
-	public int getErr() {
+	public double getErr() {
 		return err;
 	}
 	public void setErr(int err) {
 		this.err = err;
 	}
-	public int getDrops() {
+	public double getDrops() {
 		return drops;
 	}
 	public void setDrops(int drops) {
